Generate list of all kubeconfigs in kubeconfig::list
