+++ src/kubeconfig/mod.rs
@@ -4,8 +4,10 @@ use std::{fs::File, path::Path};
 use url::Url;
 
 mod import;
+mod list;
 
 pub use import::import;
+pub use list::list;
 
 #[cfg(test)]
 mod tests;
